requirements.txt
openpyxl>=3.1.0
redis>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.25.0
mcp>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
